feat: cache state using remember and add screenshots for person content
