+++ feature/details/src/main/kotlin/com/divinelink/feature/details/person/ui/PersonScreen.kt
@@ -9,7 +9,6 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -19,6 +18,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.divinelink.core.navigation.arguments.DetailsNavArguments
 import com.divinelink.core.navigation.arguments.PersonNavArguments
 import com.divinelink.core.ui.TestTags
@@ -48,7 +48,7 @@ fun PersonScreen(
   navigator: DestinationsNavigator,
   viewModel: PersonViewModel = koinViewModel(),
 ) {
-  val uiState = viewModel.uiState.collectAsState().value
+  val uiState = viewModel.uiState.collectAsStateWithLifecycle().value
   var showDropdownMenu by remember { mutableStateOf(false) }
 
   val scope = rememberCoroutineScope()
@@ -59,10 +59,17 @@ fun PersonScreen(
     derivedStateOf { connection.currentSize < 1.dp }
   }
 
-  val aboutForm = remember(uiState) {
-    uiState.forms.getOrElse(PersonTab.ABOUT.order) { null } as? PersonForm.About
+  val personDetails by remember(uiState.aboutForm) {
+    derivedStateOf { uiState.aboutForm?.personDetails }
   }
-  val personDetails = aboutForm?.personDetails?.takeIf { it is PersonDetailsUiState.Data }
+
+  val movies by remember(uiState.filteredCredits[PersonTab.MOVIES.order]) {
+    derivedStateOf { uiState.filteredCredits[PersonTab.MOVIES.order] ?: emptyMap() }
+  }
+
+  val tvShows by remember(uiState.filteredCredits[PersonTab.TV_SHOWS.order]) {
+    derivedStateOf { uiState.filteredCredits[PersonTab.TV_SHOWS.order] ?: emptyMap() }
+  }
 
   AppScaffold(
     topBar = { scrollBehavior, topAppBarColors ->
@@ -70,7 +77,9 @@ fun PersonScreen(
         AppTopAppBar(
           scrollBehavior = scrollBehavior,
           topAppBarColors = topAppBarColors,
-          text = UIText.StringText(personDetails.personDetails.person.name),
+          text = UIText.StringText(
+            (personDetails as PersonDetailsUiState.Data).personDetails.person.name,
+          ),
           isVisible = isAppBarVisible.value,
           onNavigateUp = navigator::navigateUp,
           actions = {
@@ -82,7 +91,7 @@ fun PersonScreen(
             }
 
             PersonDropdownMenu(
-              person = personDetails.personDetails.person,
+              person = (personDetails as PersonDetailsUiState.Data).personDetails.person,
               expanded = showDropdownMenu,
               onDismissDropdown = { showDropdownMenu = false },
             )
@@ -115,9 +124,8 @@ fun PersonScreen(
         onTabSelected = viewModel::onTabSelected,
         onUpdateLayoutStyle = viewModel::onUpdateLayoutStyle,
         onApplyFilter = viewModel::onApplyFilter,
-        personDetails = personDetails as PersonDetailsUiState.Data,
-        movies = uiState.filteredCredits[PersonTab.MOVIES.order] ?: emptyMap(),
-        tvShows = uiState.filteredCredits[PersonTab.TV_SHOWS.order] ?: emptyMap(),
+        movies = movies,
+        tvShows = tvShows,
       )
     }
   }
