+++ feature/details/src/main/res/values/strings.xml
@@ -40,6 +40,7 @@
   <string name="feature_details_person_tab_movies">Movies</string>
   <string name="feature_details_person_tab_tv_shows">TV Shows</string>
 
+  <string name="feature_details_person_credits_department">Department</string>
   <string name="feature_details_person_credits_department_size">%s (%s)</string>
 
   <string name="feature_details_filter_sort_release_date">Release date</string>
